Scope tracing lifecycle to cloud-agent integration
